Remove discarded components, add new contract components
