--- packages/ui/src/components/popcorn/ContractStats/index.tsx
@@ -1,99 +1,0 @@
-/* This example requires Tailwind CSS v2.0+ */
-import { CloudIcon, TrendingUpIcon } from '@heroicons/react/outline';
-import { ArrowSmDownIcon, ArrowSmUpIcon } from '@heroicons/react/solid';
-import { getDummyEmissionData } from '../recharts/dummyEmissionsData';
-import { EmissionsLineBarChart } from '../recharts/LineBarChart';
-
-const stats = [
-  {
-    id: 1,
-    name: 'co2Emissions',
-    stat: '71kg',
-    icon: CloudIcon,
-    change: '12.38%',
-    changeType: 'increase',
-  },
-  {
-    id: 2,
-    name: 'Transactions',
-    stat: '23k',
-    icon: TrendingUpIcon,
-    change: '5.4%',
-    changeType: 'increase',
-  },
-];
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
-
-export const ContractStats = () => {
-  return (
-    <div className="pb-4">
-      <dl className="mt-5 grid grid-cols-1 gap-5 sm:grid-cols-2 lg:grid-cols-4 mx-4">
-        <div className="relative bg-gray-50 pt-5 px-6 pb-6 sm:pt-6 sm:px-6 overflow-hidden">
-          <p className="my-auto h-24 text-2xl font-semibold text-gray-900 ">
-            Popcorn HYSI Staking Pool
-          </p>
-        </div>
-        {stats.map((item) => (
-          <div
-            key={item.id}
-            className="relative h-24 bg-white pt-5 px-6 pb-6 sm:pt-6 sm:px-6 shadow rounded-lg overflow-hidden"
-          >
-            <dt>
-              <div className="absolute bg-indigo-500 rounded-md p-3">
-                <item.icon className="h-6 w-6 text-white" aria-hidden="true" />
-              </div>
-            </dt>
-
-            <dt>
-              <div className="absolute bg-indigo-500 rounded-md p-3">
-                <item.icon className="h-6 w-6 text-white" aria-hidden="true" />
-              </div>
-              <p className="ml-16 text-sm font-medium text-gray-500 truncate">
-                {item.name}
-              </p>
-            </dt>
-            <dd className="ml-16  flex items-baseline ">
-              <p className="text-2xl font-semibold text-gray-900">
-                {item.stat}
-              </p>
-              <p
-                className={classNames(
-                  item.changeType === 'increase'
-                    ? 'text-green-600'
-                    : 'text-red-600',
-                  'ml-2 flex items-baseline text-sm font-semibold',
-                )}
-              >
-                {item.changeType === 'increase' ? (
-                  <ArrowSmUpIcon
-                    className="self-center flex-shrink-0 h-5 w-5 text-green-500"
-                    aria-hidden="true"
-                  />
-                ) : (
-                  <ArrowSmDownIcon
-                    className="self-center flex-shrink-0 h-5 w-5 text-red-500"
-                    aria-hidden="true"
-                  />
-                )}
-
-                <span className="sr-only">
-                  {item.changeType === 'increase' ? 'Increased' : 'Decreased'}{' '}
-                  by
-                </span>
-                {item.change}
-              </p>
-            </dd>
-          </div>
-        ))}
-        <EmissionsLineBarChart
-          data={getDummyEmissionData()}
-          width={300}
-          height={200}
-        />
-      </dl>
-    </div>
-  );
-};
