+++ packages/ui/src/components/popcorn/ContractStatsRow/index.tsx
@@ -1,8 +1,4 @@
-import {
-  CursorClickIcon,
-  MailOpenIcon,
-  UsersIcon,
-} from '@heroicons/react/solid';
+import { MailOpenIcon, UsersIcon } from '@heroicons/react/solid';
 import React from 'react';
 import { AreaBarChart } from '../recharts/AreaBarChart';
 import { getDummyEmissionData } from '../recharts/dummyEmissionsData';
@@ -25,24 +21,16 @@ const stats = [
     change: '5.4%',
     changeType: 'increase',
   },
-  {
-    id: 3,
-    name: 'Average Gas Price',
-    stat: '24.57%',
-    icon: CursorClickIcon,
-    change: '3.2%',
-    changeType: 'decrease',
-  },
 ];
 
-export const TotalStatsStackedRows = ({ emissionSummaryStats }) => {
+export const ContractStatsRow = ({ emissionSummaryStats, contractName }) => {
   return (
     <div className="py-10 mx-8">
       <div className="max-w-7xl">
         <div className="mt-2 mb-8">
           <dt>
             <h1 className="text-3xl font-bold leading-tight text-gray-900">
-              Total Stats
+              {contractName}
             </h1>
           </dt>
           <dd className=" text-base text-gray-500">
