+++ packages/ui/src/components/popcorn/StatsRow/index.stories.tsx
@@ -1,5 +1,9 @@
-// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
-import { CloudIcon } from '@heroicons/react/solid';
+import {
+  CloudIcon,
+  CursorClickIcon,
+  MailOpenIcon,
+  UsersIcon,
+} from '@heroicons/react/solid';
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
 import { EmissionSummaryStats } from '../TotalStats';
@@ -30,5 +34,30 @@ const Template: Story = (args) => <StatsRow {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {
-  item: dummyItem,
+  stats: [
+    {
+      id: 1,
+      name: 'CO2 Emissions (kg)',
+      stat: '71,897',
+      icon: UsersIcon,
+      change: '122',
+      changeType: 'increase',
+    },
+    {
+      id: 2,
+      name: 'Transactions',
+      stat: '58.16%',
+      icon: MailOpenIcon,
+      change: '5.4%',
+      changeType: 'increase',
+    },
+    {
+      id: 3,
+      name: 'Average Gas Price',
+      stat: '24.57%',
+      icon: CursorClickIcon,
+      change: '3.2%',
+      changeType: 'decrease',
+    },
+  ],
 };
