--- packages/ui/src/components/popcorn/TotalStats/index.tsx
@@ -1,86 +1,0 @@
-import { ArrowSmDownIcon, ArrowSmUpIcon } from '@heroicons/react/solid';
-import { SVGProps } from 'react';
-import { EmissionsAreaChart } from '../recharts/AreaChart';
-import { getDummyEmissionData } from '../recharts/dummyEmissionsData';
-
-export interface EmissionSummaryStats {
-  id: number;
-  name: string;
-  stat: string;
-  icon: (props: SVGProps<SVGSVGElement>) => JSX.Element;
-  change: string;
-  changeType: 'increase' | 'decrease';
-}
-
-interface TotalStatsProps {
-  emissionSummaryStats: EmissionSummaryStats[];
-}
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
-
-export const TotalStats: React.FC<TotalStatsProps> = ({
-  emissionSummaryStats,
-}) => {
-  return (
-    <div className="pb-8 bg-gray-50">
-      <h3 className="text-lg leading-6 font-medium text-gray-900 mt-16">
-        Totals between 12/04/2021 and 30/05/2021
-      </h3>
-      <dl className="mt-5 grid grid-cols-1 gap-5 sm:grid-cols-2 lg:grid-cols-4 mx-4">
-        {emissionSummaryStats.map((item) => (
-          <div
-            key={item.id}
-            className="relative h-24 bg-white pt-5 px-6 pb-6 sm:pt-6 sm:px-6 shadow rounded-lg overflow-hidden "
-          >
-            <dt>
-              <div className="absolute bg-indigo-500 rounded-md p-3">
-                <item.icon className="h-6 w-6 text-white" aria-hidden="true" />
-              </div>
-              <p className="ml-16 text-sm font-medium text-gray-500 truncate">
-                {item.name}
-              </p>
-            </dt>
-            <dd className="ml-16  flex items-baseline ">
-              <p className="text-2xl font-semibold text-gray-900">
-                {item.stat}
-              </p>
-              <p
-                className={classNames(
-                  item.changeType === 'increase'
-                    ? 'text-green-600'
-                    : 'text-red-600',
-                  'ml-2 flex items-baseline text-sm font-semibold',
-                )}
-              >
-                {item.changeType === 'increase' ? (
-                  <ArrowSmUpIcon
-                    className="self-center flex-shrink-0 h-5 w-5 text-green-500"
-                    aria-hidden="true"
-                  />
-                ) : (
-                  <ArrowSmDownIcon
-                    className="self-center flex-shrink-0 h-5 w-5 text-red-500"
-                    aria-hidden="true"
-                  />
-                )}
-
-                <span className="sr-only">
-                  {item.changeType === 'increase' ? 'Increased' : 'Decreased'}{' '}
-                  by
-                </span>
-                {item.change}
-              </p>
-            </dd>
-          </div>
-        ))}
-        <EmissionsAreaChart
-          data={getDummyEmissionData()}
-          width={300}
-          height={200}
-        />
-      </dl>
-    </div>
-  );
-};
