--- packages/ui/src/components/popcorn/recharts/AreaChart/index.tsx
@@ -1,69 +1,0 @@
-import React from 'react';
-import {
-  Area,
-  AreaChart,
-  CartesianGrid,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from 'recharts';
-
-export interface ChartData {
-  date: string;
-  co2Emissions: number;
-  numTransactions: number;
-}
-
-export interface AreaChartProps {
-  data: ChartData[];
-  height?: number;
-  width?: number;
-}
-
-export const EmissionsAreaChart: React.FC<AreaChartProps> = ({
-  data,
-  height,
-  width,
-}) => {
-  height = height || 200;
-  width = width || 300;
-  return (
-    <div>
-      <AreaChart
-        width={width}
-        height={height}
-        data={data}
-        margin={{ top: 10, right: 30, left: 0, bottom: 0 }}
-      >
-        <defs>
-          <linearGradient id="colorUv" x1="0" y1="0" x2="0" y2="1">
-            <stop offset="5%" stopColor="#818CF8" stopOpacity={0.8} />
-            <stop offset="95%" stopColor="#818CF8" stopOpacity={0} />
-          </linearGradient>
-          <linearGradient id="colorPv" x1="0" y1="0" x2="0" y2="1">
-            <stop offset="5%" stopColor="#34D399" stopOpacity={0.8} />
-            <stop offset="95%" stopColor="#34D399" stopOpacity={0} />
-          </linearGradient>
-        </defs>
-        <XAxis dataKey="date" />
-        <YAxis />
-        <CartesianGrid strokeDasharray="3 3" />
-        <Tooltip />
-        <Area
-          type="monotone"
-          dataKey="co2Emissions"
-          stroke="#818CF8"
-          fillOpacity={1}
-          fill="url(#colorUv)"
-        />
-        <Area
-          type="monotone"
-          dataKey="numTransactions"
-          stroke="#34D399"
-          fillOpacity={1}
-          fill="url(#colorPv)"
-        />
-      </AreaChart>
-    </div>
-  );
-};
